Trying to upgrade kernel under EL6 derivitives
Often the kernel-devel package for the running kernel will be missing
from the repos if the running kernel is out of date.

This attempts to upgrade the kernel if kernel-devel is not available at
the right version.

Currently it *does* work, but results in the VM getting the
VBoxGuestAdditions iso copied to the VM and installing them twice - which may
be something to do with the forced reboot for the kernel upgrade.